Extracted each pipeline (questions/answers/photos) to their own files and housed them in `./db/utils/etl`. Updated get-rows to count-rows and confirmed transform pipeline returns files with the same number of records as the original csvs
