+++ db/utils/etl/package.json
@@ -1,7 +1,7 @@
 {
   "name": "utils",
   "version": "1.0.0",
-  "main": "etl.js",
+  "main": "questions-etl.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1"
   },
